+++ prot/udp/udp.go
@@ -1,9 +1,9 @@
 package udp
 
 type UdpMsg struct {
-    SrcIP   string
-    DstIP   string
-    SrcPort uint16
-    DstPort uint16
-    Payload []byte
+	SrcIP   string
+	DstIP   string
+	SrcPort int
+	DstPort int
+	Payload []byte
 }
